update flux cache
