+++ angelslim/compressor/cache/__init__.py
@@ -13,6 +13,5 @@
 # limitations under the License.
 
 from .cache import Cache  # noqa: F401
-from .cachepipeline import FluxCachePipeline  # noqa: F401
 from .deepcache import FluxDeepCacheHelper  # noqa: F401
 from .teacache import TeaCache  # noqa: F401
